fixed small bug when classpath has some strange formatting

+++ core/src/main/scala/spark/MesosScheduler.scala
@@ -253,9 +253,11 @@ extends MScheduler with spark.Scheduler with Logging
     // Copy each JAR to a unique filename in the jarDir
     for ((path, index) <- sc.jars.zipWithIndex) {
       val file = new File(path)
-      val filename = index + "_" + file.getName
-      copyFile(file, new File(jarDir, filename))
-      filenames += filename
+      if (file.exists) {
+        val filename = index + "_" + file.getName
+        copyFile(file, new File(jarDir, filename))
+        filenames += filename
+      }
     }
     // Create the server
     jarServer = new HttpServer(jarDir)
